feat: Delete CNTKand related utils (#1743)
* Delete CNTKand related utils

* added UnrollImage back

* build fix

* fix dotnet tests
