+++ core/src/main/scala/com/microsoft/azure/synapse/ml/image/UnrollImage.scala
@@ -21,6 +21,7 @@ import org.apache.spark.sql.{DataFrame, Dataset, Row}
 import java.awt.Color
 import java.awt.color.ColorSpace
 import java.awt.image.BufferedImage
+import java.awt.{Image => JImage}
 import scala.math.round
 
 object UnrollImage extends DefaultParamsReadable[UnrollImage] {
@@ -132,11 +133,28 @@ object UnrollImage extends DefaultParamsReadable[UnrollImage] {
     val biOpt = ImageUtils.safeRead(bytes)
     biOpt.map { bi =>
       (height, width) match {
-        case (Some(h), Some(w)) => unrollBI(ResizeUtils.resizeBufferedImage(w, h, nChannels)(bi))
+        case (Some(h), Some(w)) => unrollBI(resizeBufferedImage(w, h, nChannels)(bi))
         case (None, None) => unrollBI(bi)
         case _ =>
           throw new IllegalArgumentException("Height and width must either both be specified or unspecified")
       }
+    }
+  }
+
+  private[ml] def resizeBufferedImage(width: Int,
+                                      height: Int,
+                                      channels: Option[Int])(image: BufferedImage): BufferedImage = {
+    val imgType = channels.map(ImageUtils.channelsToType).getOrElse(image.getType)
+
+    if (image.getWidth == width && image.getHeight == height && image.getType == imgType) {
+      image
+    } else {
+      val resizedImage = image.getScaledInstance(width, height, JImage.SCALE_DEFAULT)
+      val bufferedImage = new BufferedImage(width, height, imgType)
+      val g = bufferedImage.createGraphics()
+      g.drawImage(resizedImage, 0, 0, null) //scalastyle:ignore null
+      g.dispose()
+      bufferedImage
     }
   }
 
@@ -148,7 +166,6 @@ object UnrollImage extends DefaultParamsReadable[UnrollImage] {
   *
   * @param uid The id of the module
   */
-@deprecated("Please use 'OnnxModel'.", since="0.10.2")
 class UnrollImage(val uid: String) extends Transformer
   with HasInputCol with HasOutputCol with Wrappable with DefaultParamsWritable with BasicLogging {
   logClass()
