+++ core/src/test/scala/com/microsoft/azure/synapse/ml/core/test/fuzzing/Fuzzing.scala
@@ -191,8 +191,7 @@ trait DotnetTestFuzzing[S <: PipelineStage] extends TestBase with DataFrameEqual
       .split(".".toCharArray).map(capitalize).mkString(".")
     val externalLoaderImports = conf.name match {
       case "synapseml-deep-learning" =>
-        s"""using Synapse.ML.Cntk;
-           |using Synapse.ML.Onnx;
+        s"""using Synapse.ML.Onnx;
            |using Synapse.ML.Stages;
            |""".stripMargin
       case _ => ""
